v050825

+++ app/src/defines.h
@@ -284,42 +284,42 @@ typedef union data128 {
 
   void* address[2];
 
-   data128(void) {
+  data128(void) {
     this->u64[0] = 0u;
     this->u64[1] = 0u;
   }
   
-   data128(::i64 value1, ::i64 value2 = 0) {
+  data128(::i64 value1, ::i64 value2 = 0) {
     this->i64[0] = value1;
     this->i64[1] = value2;
   }
-   data128(::u64 value1, ::u64 value2 = 0){
+  data128(::u64 value1, ::u64 value2 = 0){
     this->u64[0] = value1;
     this->u64[1] = value2;
   }
-   data128(::f64 value1, ::f64 value2 = 0){
+  data128(::f64 value1, ::f64 value2 = 0){
     this->f64[0] = value1;
     this->f64[1] = value2;
   }
-   data128(::i32 value1, ::i32 value2 = 0, ::i32 value3 = 0, ::i32 value4 = 0){
+  data128(::i32 value1, ::i32 value2 = 0, ::i32 value3 = 0, ::i32 value4 = 0){
     this->i32[0] = value1;
     this->i32[1] = value2;
     this->i32[2] = value3;
     this->i32[3] = value4;
   }
-   data128(::u32 value1, ::u32 value2 = 0, ::u32 value3 = 0, ::u32 value4 = 0){
+  data128(::u32 value1, ::u32 value2 = 0, ::u32 value3 = 0, ::u32 value4 = 0){
     this->u32[0] = value1;
     this->u32[1] = value2;
     this->u32[2] = value3;
     this->u32[3] = value4;
   }
-   data128(::f32 value1, ::f32 value2 = 0, ::f32 value3 = 0, ::f32 value4 = 0){
+  data128(::f32 value1, ::f32 value2 = 0, ::f32 value3 = 0, ::f32 value4 = 0){
     this->f32[0] = value1;
     this->f32[1] = value2;
     this->f32[2] = value3;
     this->f32[3] = value4;
   }
-   data128(::u16 value1, ::u16 value2 = 0, ::u16 value3 = 0, ::u16 value4 = 0, ::u16 value5 = 0, ::u16 value6 = 0, ::u16 value7 = 0, ::u16 value8 = 0){
+  data128(::u16 value1, ::u16 value2 = 0, ::u16 value3 = 0, ::u16 value4 = 0, ::u16 value5 = 0, ::u16 value6 = 0, ::u16 value7 = 0, ::u16 value8 = 0){
     this->u16[0] = value1;
     this->u16[1] = value2;
     this->u16[2] = value3;
@@ -329,7 +329,7 @@ typedef union data128 {
     this->u16[6] = value7;
     this->u16[7] = value8;
   }
-   data128(::i16 value1, ::i16 value2 = 0, ::i16 value3 = 0, ::i16 value4 = 0, ::i16 value5 = 0, ::i16 value6 = 0, ::i16 value7 = 0, ::i16 value8 = 0){
+  data128(::i16 value1, ::i16 value2 = 0, ::i16 value3 = 0, ::i16 value4 = 0, ::i16 value5 = 0, ::i16 value6 = 0, ::i16 value7 = 0, ::i16 value8 = 0){
     this->i16[0] = value1;
     this->i16[1] = value2;
     this->i16[2] = value3;
@@ -339,35 +339,35 @@ typedef union data128 {
     this->i16[6] = value7;
     this->i16[7] = value8;
   }
-   data128(::u16* value, ::u16 len){
+  data128(::u16* value, ::u16 len){
     if (len < 8) {
       for (int i=0; i<len; ++i) {
         this->u16[i] = value[i];
       }
     }
   }
-   data128(::i16* value, ::u16 len){
+  data128(::i16* value, ::u16 len){
     if (len < 8) {
       for (int i=0; i<len; ++i) {
         this->i16[i] = value[i];
       }
     }
   }
-   data128(::u8* value, ::u16 len){
+  data128(::u8* value, ::u16 len){
     if (len < 16) {
       for (int i=0; i<len; ++i) {
         this->u8[i] = value[i];
       }
     }
   }
-   data128(::i8* value, ::u16 len){
+  data128(::i8* value, ::u16 len){
     if (len < 16) {
       for (int i=0; i<len; ++i) {
         this->i8[i] = value[i];
       }
     }
   }
-   data128(char* value, ::u16 len){
+  data128(char* value, ::u16 len){
     if (len < 16) {
       for (int i=0; i<len; ++i) {
         this->c[i] = value[i];
@@ -375,7 +375,7 @@ typedef union data128 {
       this->c[len-1] = '\0';
     }
   }
-   data128(const char* value, ::u16 len){
+  data128(const char* value, ::u16 len){
     if (len < 16) {
       for (int i=0; i<len; ++i) {
         this->c[i] = value[i];
@@ -383,7 +383,7 @@ typedef union data128 {
       this->c[len-1] = '\0';
     }
   }
-   data128(void* address1, void* address2 = nullptr){
+  data128(void* address1, void* address2 = nullptr){
     this->address[0] = address1;
     this->address[1] = address2;
   }
@@ -408,32 +408,32 @@ typedef union data256 {
   char c[32];
   void* address[4];
 
-   data256(void){
+  data256(void){
     this->u64[0] = 0u;
     this->u64[1] = 0u;
     this->u64[2] = 0u;
     this->u64[3] = 0u;
   }
   
-   data256(::i64 value1, ::i64 value2 = 0, ::i64 value3 = 0, ::i64 value4 = 0){
+  data256(::i64 value1, ::i64 value2 = 0, ::i64 value3 = 0, ::i64 value4 = 0){
     this->i64[0] = value1;
     this->i64[1] = value2;
     this->i64[2] = value3;
     this->i64[3] = value4;
   }
-   data256(::u64 value1, ::u64 value2 = 0, ::u64 value3 = 0, ::u64 value4 = 0){
+  data256(::u64 value1, ::u64 value2 = 0, ::u64 value3 = 0, ::u64 value4 = 0){
     this->u64[0] = value1;
     this->u64[1] = value2;
     this->u64[2] = value3;
     this->u64[3] = value4;
   }
-   data256(::f64 value1, ::f64 value2 = 0, ::f64 value3 = 0, ::f64 value4 = 0){
+  data256(::f64 value1, ::f64 value2 = 0, ::f64 value3 = 0, ::f64 value4 = 0){
     this->f64[0] = value1;
     this->f64[1] = value2;
     this->f64[2] = value3;
     this->f64[3] = value4;
   }
-   data256(::i32 value1, ::i32 value2 = 0, ::i32 value3 = 0, ::i32 value4 = 0, ::i32 value5 = 0, ::i32 value6 = 0, ::i32 value7 = 0, ::i32 value8 = 0){
+  data256(::i32 value1, ::i32 value2 = 0, ::i32 value3 = 0, ::i32 value4 = 0, ::i32 value5 = 0, ::i32 value6 = 0, ::i32 value7 = 0, ::i32 value8 = 0){
     this->i32[0] = value1;
     this->i32[1] = value2;
     this->i32[2] = value3;
@@ -443,7 +443,7 @@ typedef union data256 {
     this->i32[6] = value7;
     this->i32[7] = value8;
   }
-   data256(::u32 value1, ::u32 value2 = 0, ::u32 value3 = 0, ::u32 value4 = 0, ::u32 value5 = 0, ::u32 value6 = 0, ::u32 value7 = 0, ::u32 value8 = 0){
+  data256(::u32 value1, ::u32 value2 = 0, ::u32 value3 = 0, ::u32 value4 = 0, ::u32 value5 = 0, ::u32 value6 = 0, ::u32 value7 = 0, ::u32 value8 = 0){
     this->u32[0] = value1;
     this->u32[1] = value2;
     this->u32[2] = value3;
@@ -453,7 +453,7 @@ typedef union data256 {
     this->u32[6] = value7;
     this->u32[7] = value8;
   }
-   data256(::f32 value1, ::f32 value2 = 0, ::f32 value3 = 0, ::f32 value4 = 0, ::f32 value5 = 0, ::f32 value6 = 0, ::f32 value7 = 0, ::f32 value8 = 0){
+  data256(::f32 value1, ::f32 value2 = 0, ::f32 value3 = 0, ::f32 value4 = 0, ::f32 value5 = 0, ::f32 value6 = 0, ::f32 value7 = 0, ::f32 value8 = 0){
     this->f32[0] = value1;
     this->f32[1] = value2;
     this->f32[2] = value3;
@@ -463,35 +463,35 @@ typedef union data256 {
     this->f32[6] = value7;
     this->f32[7] = value8;
   }
-   data256(::u16* value, ::u16 len){
+  data256(::u16* value, ::u16 len){
     if (len < 16) {
       for (int i=0; i<len; ++i) {
         this->u16[i] = value[i];
       }
     }
   }
-   data256(::i16* value, ::u16 len){
+  data256(::i16* value, ::u16 len){
     if (len < 16) {
       for (int i=0; i<len; ++i) {
         this->i16[i] = value[i];
       }
     }
   }
-   data256(::u8* value, ::u16 len){
+  data256(::u8* value, ::u16 len){
     if (len < 32) {
       for (int i=0; i<len; ++i) {
         this->u8[i] = value[i];
       }
     }
   }
-   data256(::i8* value, ::u16 len){
+  data256(::i8* value, ::u16 len){
     if (len < 32) {
       for (int i=0; i<len; ++i) {
         this->i8[i] = value[i];
       }
     }
   }
-   data256(char* value, ::u16 len){
+  data256(char* value, ::u16 len){
     if (len < 32) {
       for (int i=0; i<len; ++i) {
         this->c[i] = value[i];
@@ -499,7 +499,7 @@ typedef union data256 {
       this->c[len-1] = '\0';
     }
   }
-   data256(const char* value, ::u16 len){
+  data256(const char* value, ::u16 len){
     if (len < 32) {
       for (int i=0; i<len; ++i) {
         this->c[i] = value[i];
@@ -507,7 +507,7 @@ typedef union data256 {
       this->c[len-1] = '\0';
     }
   }
-   data256(void* address1, void* address2 = nullptr, void* address3 = nullptr, void* address4 = nullptr){
+  data256(void* address1, void* address2 = nullptr, void* address3 = nullptr, void* address4 = nullptr){
     this->address[0] = address1;
     this->address[1] = address2;
     this->address[2] = address3;
@@ -517,16 +517,14 @@ typedef union data256 {
 
 typedef struct data_pack {
   data_type type_flag;
-  u16 array_lenght;
-
+  i32 array_lenght;
   data128 data;
-   data_pack() {
+  data_pack() {
     this->type_flag = DATA_TYPE_UNRESERVED;
     this->array_lenght = 0;
     this->data = data128();
   };
-   data_pack(data_type type, data128 buffer, u16 len) : data_pack()
-  {
+  data_pack(data_type type, data128 buffer, i32 len) : data_pack() {
     this->type_flag = type;
     this->array_lenght = len; 
     this->data = buffer;
